form to add customised exercise and functions
